Issue-46: Remove all documents form subscribers
