+++ src/Khatovar/Bundle/DocumentsBundle/DependencyInjection/KhatovarDocumentsExtension.php
@@ -42,7 +42,6 @@ class KhatovarDocumentsExtension extends Extension
         $loader->load('entity_listeners.yml');
         $loader->load('factories.yml');
         $loader->load('form_creators.yml');
-        $loader->load('form_subscribers.yml');
         $loader->load('form_types.yml');
         $loader->load('repositories.yml');
         $loader->load('savers.yml');
