--- src/Khatovar/Bundle/DocumentsBundle/Resources/config/form_subscribers.yml
@@ -1,17 +1,0 @@
-services:
-    khatovar_documents.form.subscriber.add_folder_to_parent:
-        class: Khatovar\Bundle\DocumentsBundle\Form\Subscriber\AddFolderToParentSubscriber
-        arguments:
-            - '@translator'
-            - '@khatovar_documents.repositories.folder'
-
-    khatovar_documents.form.subscriber.add_file_to_folder:
-        class: Khatovar\Bundle\DocumentsBundle\Form\Subscriber\AddFileToFolderSubscriber
-        arguments:
-            - '@translator'
-            - '@khatovar_documents.repositories.folder'
-
-    khatovar_documents.form.subscriber.file:
-        class: Khatovar\Bundle\DocumentsBundle\Form\Subscriber\FileSubscriber
-        arguments:
-            - '@translator'
