+++ src/Khatovar/Bundle/DocumentsBundle/Resources/config/form_types.yml
@@ -9,20 +9,22 @@ services:
     khatovar_documents.form.type.file:
         class: Khatovar\Bundle\DocumentsBundle\Form\Type\FileType
         arguments:
-            - '@khatovar_documents.form.subscriber.file'
+            - '@translator'
         tags:
             - { name: form.type }
 
     khatovar_documents.form.type.move_folder:
         class: Khatovar\Bundle\DocumentsBundle\Form\Type\FolderMoveType
         arguments:
-            - '@khatovar_documents.form.subscriber.add_folder_to_parent'
+            - '@translator'
+            - '@khatovar_documents.repositories.folder'
         tags:
             - { name: form.type }
 
     khatovar_documents.form.type.move_file:
         class: Khatovar\Bundle\DocumentsBundle\Form\Type\FileMoveType
         arguments:
-            - '@khatovar_documents.form.subscriber.add_file_to_folder'
+            - '@translator'
+            - '@khatovar_documents.repositories.folder'
         tags:
             - { name: form.type }
